requirements.txt
Django==5.0.6
Flask==3.0.3
requests==2.31.0
python-dotenv
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
